refactor: create a function for data submitting
create a function for submitting user data via a fetch method

MODULE-1-PART-3-gallery
